+++ fitfunc.h
